refactor: migrate from vite to nextjs
