+++ .gitignore
@@ -23,3 +23,6 @@ dist-ssr
 *.sln
 *.sw?
 coverage
+
+.next
+next-env.d.ts
