+++ package.json
@@ -1,71 +1,32 @@
 {
-  "name": "rsschool-react",
+  "name": "my-app",
+  "version": "0.1.0",
   "private": true,
-  "version": "0.0.0",
-  "type": "module",
   "scripts": {
-    "dev": "vite",
-    "build": "tsc -b && vite build",
-    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
-    "preview": "vite preview",
-    "format:fix": "prettier --config .prettierrc 'src/**/*.ts' --write",
-    "prepare": "husky install",
-    "test": "vitest run --coverage",
-    "deploy": "vite build && gh-pages -d dist"
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint"
   },
   "dependencies": {
-    "@reduxjs/toolkit": "^2.2.6",
-    "@testing-library/react-hooks": "^8.0.1",
-    "@types/react-redux": "^7.1.33",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
+    "@reduxjs/toolkit": "^2.2.7",
+    "@vitejs/plugin-react": "^4.3.1",
+    "next": "14.2.5",
+    "react": "^18",
+    "react-dom": "^18",
     "react-redux": "^9.1.2",
-    "react-router-dom": "^6.24.1",
-    "redux": "^5.0.1"
+    "react-router-dom": "^6.25.1"
   },
   "devDependencies": {
-    "@testing-library/dom": "^10.3.2",
-    "@testing-library/jest-dom": "^6.4.6",
+    "@testing-library/dom": "^10.4.0",
+    "@testing-library/jest-dom": "^6.4.8",
     "@testing-library/react": "^16.0.0",
-    "@testing-library/user-event": "^14.5.2",
-    "@types/jest": "^29.5.12",
-    "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.3.0",
-    "@typescript-eslint/eslint-plugin": "^7.13.1",
-    "@typescript-eslint/parser": "^7.13.1",
-    "@vitejs/plugin-react": "^4.3.1",
-    "@vitest/coverage-v8": "^2.0.3",
-    "eslint": "^8.57.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.1.3",
-    "eslint-plugin-react-compiler": "^0.0.0-experimental-0998c1e-20240625",
-    "eslint-plugin-react-hooks": "^4.6.2",
-    "eslint-plugin-react-refresh": "^0.4.7",
-    "gh-pages": "^6.1.1",
-    "happy-dom": "^14.12.3",
-    "husky": "^8.0.0",
-    "identity-obj-proxy": "^3.0.0",
-    "jest": "^29.7.0",
-    "jest-environment-jsdom": "^29.7.0",
-    "jsdom": "^24.1.0",
-    "lint-staged": "^15.2.7",
-    "msw": "^2.3.2",
-    "prettier": "^3.3.2",
-    "ts-jest": "^29.2.2",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.5.3",
-    "vite": "^5.3.1",
-    "vitest": "^2.0.2"
-  },
-  "lint-staged": {
-    "**/*.{js,jsx,ts,tsx}": [
-      "npx eslint --fix "
-    ]
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "sh .husky/pre-commit",
-      "pre-push": "sh .husky/pre-push"
-    }
+    "@types/node": "^20",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
+    "eslint": "^9.8.0",
+    "redux": "^5.0.1",
+    "typescript": "^5",
+    "vitest": "^2.0.4"
   }
 }
