--- src/components/card/Card.css
@@ -1,32 +1,0 @@
-.person {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-
-    width:600px;
-    margin:10px;
-    padding: 10px;
-
-    background: white;
-    border:none;
-    border-radius: 5px;
-
-    color: black;
-}
-
-.checkbox {
-    appearance: none;
-    border:solid var(--accent-color) 2px;
-    border-radius: 50%;
-    width: 20px;
-    height: 20px;
-}
-  
-.checkbox:hover {
-    background-color: var(--hover-color);
-}
-
-.checkbox:checked {
-    background-color: var(--accent-color);
-}
-
