--- src/components/details/Details.css
@@ -1,16 +1,0 @@
-.person-header {
-    color: var(--accent-color);
-}
-
-.close-button {
-    background-color: var(--accent-color);
-    color: white;
-    border: none;
-    padding: 8px 16px;
-    text-align: center;
-    text-decoration: none;
-    display: inline-block;
-    font-size: 16px;
-    margin-top: 10px;
-    cursor: pointer;
-}
