+++ src/components/details/Details.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
-import './Details.css';
-import { useGetDetailsByIdQuery } from '../../redux/starWarsAPI';
+import { useGetDetailsByIdQuery } from '../redux/starWarsAPI';
 
 interface Props {
   id: number;
