--- src/components/list/List.css
@@ -1,13 +1,0 @@
-ul {
-    margin: 0;
-    padding: 0;
-
-    list-style-type: none;
-}
-
-.result-list {
-    width:100%;
-    display: flex;
-    flex-direction: column;
-    place-items: center;
-}
