--- src/components/search/Search.css
@@ -1,26 +1,0 @@
-.search-bar {
-    width: 50vw;
-}
-
-.search-bar > form > input {
-    width: 70%;
-    height: 30px;
-    padding: 2px;
-
-    border: none;
-    background-color: whitesmoke;
-    border-radius: 5px 0 0 5px;
-
-    color:black;
-}
-
-.search-bar > form > button {
-    width: 20%;
-    height: 34px;
-
-    color: white;
-
-    border:none;
-    background-color: var(--accent-color);
-    border-radius: 0 5px 5px 0;
-}
