--- src/error_handling/ErrorBoundary.css
@@ -1,10 +1,0 @@
-.errorMessage {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    position: fixed;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-}
