+++ src/models/useLocalStorage.ts
@@ -2,22 +2,22 @@ import React from "react";
 
 function useLocalStorage(key: string, defaultValue: string): [string, (value: string) => void] {
     const [value, setValue] = React.useState<string>(() => {
-        const storedValue = localStorage.getItem(key);
-        return storedValue ? JSON.parse(storedValue) : defaultValue;
+        if (typeof window !== "undefined") {
+            const storedValue = localStorage.getItem(key);
+            return storedValue ? JSON.parse(storedValue) : defaultValue;
+        } else {
+            return defaultValue;
+        }
     });
 
     React.useEffect(() => {
-        try {
-            localStorage.setItem(key, JSON.stringify(value));
-        } catch (error) {
-            console.error(`Error setting localStorage key "${key}":`, error);
+        if (typeof window !== "undefined") {
+            try {
+                localStorage.setItem(key, JSON.stringify(value));
+            } catch (error) {
+                console.error(`Error setting localStorage key "${key}":`, error);
+            }
         }
-    }, [key, value]);
-
-    React.useEffect(() => {
-        return () => {
-            localStorage.setItem(key, JSON.stringify(value));
-        };
     }, [key, value]);
 
     return [value, setValue];
