+++ src/pages/index.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import App from '../components/home/App';
+
+const HomePage = () => {
+  return <App></App>
+};
+
+export default HomePage;
