+++ src/test/Card.test.tsx
@@ -5,7 +5,7 @@ import { fireEvent, render, screen, waitFor } from "@testing-library/react";
 import { mockPeople } from "../mocks/mockPeople";
 import { vi } from "vitest";
 import { Provider } from "react-redux";
-import store from "../redux/store";
+import store from "../components/redux/store";
 
 const mockPerson = mockPeople[0];
